fix: tul adjustments

+++ Core/Consts/ExerciseConsts.cs
@@ -6,7 +6,7 @@ public static class ExerciseConsts
     /// <summary>
     /// ~24 per exercise: 6reps * 4sets; 8reps * 3sets; 12reps * 2sets; 60s total TUT / 2.5.
     /// </summary>
-    public const double TargetVolumePerExercise = 24;
+    public const int TargetVolumePerExercise = 24;
 
     /// <summary>
     /// How many days until exercises become stale if they haven't been able to be seen.
